+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/auth/jwt/JWTFilter.java
@@ -9,7 +9,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import com.ssafy.shinhanflow.auth.custom.CustomUserDetails;
-import com.ssafy.shinhanflow.entity.MemberEntity;
+import com.ssafy.shinhanflow.domain.entity.MemberEntity;
 import com.ssafy.shinhanflow.repository.MemberRepository;
 
 import io.jsonwebtoken.ExpiredJwtException;
