+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/domain/entity/TriggerEntity.java
@@ -1,4 +1,4 @@
-package com.ssafy.shinhanflow.entity;
+package com.ssafy.shinhanflow.domain.entity;
 
 import java.time.LocalDateTime;
 
@@ -6,8 +6,6 @@ import org.hibernate.annotations.ColumnDefault;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -54,10 +52,6 @@ public class TriggerEntity {
 	@ColumnDefault("NULL")
 	@Column(name = "deleted_at", insertable = false, columnDefinition = "TIMESTAMP")
 	private LocalDateTime deletedAt;
-
-	public enum Category {
-		TIME, PRODUCT, TRANSFER, EXCHANGE
-	}
 
 	@Builder
 	public TriggerEntity(Long flowId, Long memberId, int code, String data) {
