+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/domain/trigger/account/TransferTrigger.java
@@ -1,9 +1,9 @@
-package com.ssafy.shinhanflow.flow.trigger.account;
+package com.ssafy.shinhanflow.domain.trigger.account;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import com.ssafy.shinhanflow.flow.trigger.Trigger;
+import com.ssafy.shinhanflow.domain.trigger.Trigger;
 
 import lombok.NonNull;
 
@@ -14,8 +14,8 @@ public record TransferTrigger(@NonNull String fromAccount, @NonNull String toAcc
 	public boolean isTriggered() {
 		//todo: 송금하면 이니까 fromAccount에서 최근 거래내역 가져오기
 		List<Long> amounts = new ArrayList<>();
-		for(Long amount: amounts){
-			if(amount.compareTo(this.amount)  >= 0){
+		for (Long amount : amounts) {
+			if (amount.compareTo(this.amount) >= 0) {
 				return true;
 			}
 		}
