+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/domain/trigger/date/DayOfWeekTrigger.java
@@ -1,10 +1,10 @@
-package com.ssafy.shinhanflow.flow.trigger.date;
+package com.ssafy.shinhanflow.domain.trigger.date;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.util.Set;
 
-import com.ssafy.shinhanflow.flow.trigger.Trigger;
+import com.ssafy.shinhanflow.domain.trigger.Trigger;
 
 import lombok.NonNull;
 
@@ -14,8 +14,8 @@ public record DayOfWeekTrigger(@NonNull Set<DayOfWeek> daysOfWeek) implements Tr
 		LocalDate today = LocalDate.now();
 		DayOfWeek todayOfWeek = today.getDayOfWeek();
 
-		for(DayOfWeek day: daysOfWeek){
-			if (todayOfWeek.equals(day)){
+		for (DayOfWeek day : daysOfWeek) {
+			if (todayOfWeek.equals(day)) {
 				return true;
 			}
