+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/domain/trigger/date/MultiDateTrigger.java
@@ -1,9 +1,9 @@
-package com.ssafy.shinhanflow.flow.trigger.date;
+package com.ssafy.shinhanflow.domain.trigger.date;
 
 import java.time.LocalDate;
 import java.util.Set;
 
-import com.ssafy.shinhanflow.flow.trigger.Trigger;
+import com.ssafy.shinhanflow.domain.trigger.Trigger;
 
 import lombok.NonNull;
 
@@ -11,8 +11,8 @@ public record MultiDateTrigger(@NonNull Set<LocalDate> dates) implements Trigger
 	@Override
 	public boolean isTriggered() {
 		LocalDate today = LocalDate.now();
-		for(LocalDate date: dates){
-			if(today.isEqual(date)){
+		for (LocalDate date : dates) {
+			if (today.isEqual(date)) {
 				return true;
 			}
 		}
