--- be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/flow/action/notification/ExchangeRateNotificationAction.java
@@ -1,20 +1,0 @@
-package com.ssafy.shinhanflow.flow.action.notification;
-
-import com.ssafy.shinhanflow.flow.action.Action;
-import com.ssafy.shinhanflow.flow.trigger.exchange.ExchangeRateTrigger.Currency;
-
-import lombok.Builder;
-
-@Builder
-public record ExchangeRateNotificationAction(
-	Long memberId,
-	Long flowId,
-	Currency currency
-
-) implements Action {
-
-	@Override
-	public boolean execute() {
-		return false;
-	}
-}
