+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/ActionRepository.java
@@ -2,7 +2,7 @@ package com.ssafy.shinhanflow.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ssafy.shinhanflow.entity.ActionEntity;
+import com.ssafy.shinhanflow.domain.entity.ActionEntity;
 
 public interface ActionRepository extends JpaRepository<ActionEntity, Long> {
 }
