+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/FlowRepository.java
@@ -2,7 +2,7 @@ package com.ssafy.shinhanflow.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ssafy.shinhanflow.entity.FlowEntity;
+import com.ssafy.shinhanflow.domain.entity.FlowEntity;
 
 public interface FlowRepository extends JpaRepository<FlowEntity, Long> {
 }
