+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package com.ssafy.shinhanflow.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ssafy.shinhanflow.entity.MemberEntity;
+import com.ssafy.shinhanflow.domain.entity.MemberEntity;
 
 @Repository
 public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
