+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/repository/TriggerRepository.java
@@ -2,7 +2,7 @@ package com.ssafy.shinhanflow.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ssafy.shinhanflow.entity.TriggerEntity;
+import com.ssafy.shinhanflow.domain.entity.TriggerEntity;
 
 public interface TriggerRepository extends JpaRepository<TriggerEntity, Long> {
 }
